agregue una estructutra para envio de datos con 3 campos: tipo de proceso, tipo de mensaje y mensaje (este ultimo habra q ver de reemplazarlo dps con un tipo file o algo así, dependiendo de lo qu emandemos) lo implemente en memoria que recibe y kernel y cpu que envían. lo implemente igual asi que dps viendo lo que necesitamos de cada uno y si lo metemos en el resto de los procesos hay que ver de no repetir ese codigo

+++ Consola/consolaConfigurators.h
@@ -11,7 +11,7 @@ typedef struct {
 
 	char *ip_kernel;
 	char *puerto_kernel;
-	uint8_t tipo_de_proceso;
+	uint32_t tipo_de_proceso;
 } tConsola;
 
 /* Carga los datos de configuracion en una estructura.
